refactor(extension): extract conversation tabs and message composer siblings
